+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - #2734 Fixed Client restricted categories in the Sample add view form
+- #2739 Fix snapshot versioning
 - #2737 Fix form values omitted on initial load for UID reference fields in sample add form
 - #2735 Fix API field validation for DX contents
 - #2736 Fix profile services are not applied when creating samples programmatically